+++ .storybook/stories/AuthenticatedAppShell/icons.tsx
@@ -1,6 +1,24 @@
 import { createIcon } from '@ag.ds-next/react/icon';
 import { Fragment } from 'react';
 
+export const ChartPieIcon = createIcon(
+	<Fragment>
+		<path d="M12 22C17.5228 22 22 17.5228 22 12C22 6.47715 17.5228 2 12 2C6.47715 2 2 6.47715 2 12C2 17.5228 6.47715 22 12 22Z" />
+		<path d="M7 4L12 12" />
+		<path d="M17.4545 20.1818L12 12" />
+		<path d="M22 12H12" />
+	</Fragment>,
+	'ChartPieIcon'
+);
+
+export const ChevronsLeftIcon = createIcon(
+	<Fragment>
+		<path d="M18.5 7.5L14.1179 11.8821C14.0528 11.9472 14.0528 12.0528 14.1179 12.1179L18.5 16.5" />
+		<path d="M9.5 7.5L5.11785 11.8821C5.05276 11.9472 5.05276 12.0528 5.11785 12.1179L9.5 16.5" />
+	</Fragment>,
+	'ChevronsLeftIcon'
+);
+
 export const CogIcon = createIcon(
 	<Fragment>
 		<path d="M10.7063 3.49748C10.8695 3.67865 11.0689 3.82351 11.2916 3.92267C11.5144 4.02183 11.7555 4.07306 11.9993 4.07306C12.2431 4.07306 12.4842 4.02183 12.707 3.92267C12.9297 3.82351 13.1291 3.67865 13.2923 3.49748L14.1336 2.57425C14.3747 2.30737 14.6925 2.12166 15.0433 2.0426C15.3942 1.96355 15.7609 1.99503 16.0931 2.13272C16.4254 2.27041 16.7069 2.50757 16.899 2.81164C17.0911 3.11571 17.1844 3.47179 17.1661 3.83101L17.103 5.07621C17.0908 5.31912 17.1296 5.5619 17.2169 5.7889C17.3042 6.0159 17.438 6.22209 17.6099 6.3942C17.7817 6.5663 17.9877 6.7005 18.2145 6.78815C18.4413 6.8758 18.684 6.91496 18.9269 6.9031L20.1719 6.83995C20.5309 6.82225 20.8865 6.91601 21.1901 7.10837C21.4937 7.30073 21.7303 7.58229 21.8676 7.91444C22.005 8.2466 22.0362 8.6131 21.9571 8.96371C21.878 9.31433 21.6925 9.6319 21.4258 9.87291L20.4992 10.709C20.3183 10.8724 20.1736 11.072 20.0746 11.2948C19.9757 11.5177 19.9245 11.7588 19.9245 12.0026C19.9245 12.2465 19.9757 12.4876 20.0746 12.7104C20.1736 12.9333 20.3183 13.1329 20.4992 13.2963L21.4258 14.1324C21.6927 14.3735 21.8784 14.6914 21.9574 15.0422C22.0364 15.3931 22.005 15.7599 21.8673 16.0922C21.7296 16.4245 21.4925 16.706 21.1885 16.8982C20.8845 17.0903 20.5284 17.1836 20.1693 17.1653L18.9242 17.1022C18.6808 17.0897 18.4375 17.1284 18.21 17.2159C17.9825 17.3033 17.7759 17.4376 17.6035 17.61C17.4312 17.7823 17.297 17.989 17.2095 18.2165C17.1221 18.444 17.0833 18.6874 17.0959 18.9309L17.159 20.1761C17.1749 20.5335 17.0804 20.8871 16.8884 21.189C16.6964 21.4908 16.4161 21.7262 16.0856 21.8632C15.7552 22.0002 15.3905 22.0321 15.0413 21.9546C14.6921 21.8772 14.3752 21.694 14.1336 21.4301L13.2968 20.5042C13.1334 20.3233 12.934 20.1787 12.7112 20.0797C12.4885 19.9807 12.2475 19.9295 12.0037 19.9295C11.76 19.9295 11.519 19.9807 11.2963 20.0797C11.0735 20.1787 10.874 20.3233 10.7107 20.5042L9.87032 21.4301C9.62916 21.6953 9.3121 21.8796 8.96239 21.9579C8.61268 22.0362 8.24732 22.0048 7.91613 21.8678C7.58494 21.7309 7.30403 21.4952 7.1117 21.1927C6.91937 20.8903 6.82497 20.5359 6.84138 20.1778L6.90541 18.9326C6.91793 18.6892 6.87921 18.4458 6.79175 18.2183C6.7043 17.9907 6.57006 17.7841 6.39771 17.6117C6.22537 17.4394 6.01876 17.3051 5.79126 17.2176C5.56376 17.1302 5.32043 17.0914 5.07702 17.104L3.832 17.1671C3.47298 17.1859 3.11696 17.093 2.81281 16.9013C2.50867 16.7096 2.27132 16.4284 2.13334 16.0963C1.99537 15.7643 1.96353 15.3977 2.04221 15.0469C2.12089 14.696 2.30622 14.3781 2.57276 14.1368L3.49852 13.3008C3.67944 13.1373 3.82407 12.9377 3.92306 12.7149C4.02205 12.4921 4.07319 12.2509 4.07319 12.0071C4.07319 11.7633 4.02205 11.5221 3.92306 11.2993C3.82407 11.0765 3.67944 10.8769 3.49852 10.7134L2.57276 9.87291C2.30708 9.63187 2.12232 9.31466 2.04372 8.96462C1.96512 8.61459 1.99653 8.24882 2.13364 7.91731C2.27076 7.58581 2.5069 7.30474 2.80978 7.11254C3.11265 6.92034 3.46749 6.82638 3.82578 6.84351L5.07079 6.90666C5.31466 6.91948 5.5585 6.88086 5.78648 6.79329C6.01445 6.70573 6.22146 6.57118 6.39405 6.39839C6.56665 6.2256 6.70098 6.01843 6.78831 5.79034C6.87564 5.56225 6.91402 5.31833 6.90096 5.07444L6.84138 3.82835C6.82458 3.47014 6.91869 3.11548 7.11088 2.81274C7.30306 2.51001 7.58396 2.27396 7.91524 2.1368C8.24652 1.99963 8.61205 1.96804 8.96194 2.04633C9.31184 2.12462 9.62906 2.30898 9.87032 2.57425L10.7063 3.49748Z" />
@@ -18,10 +36,19 @@ export const ExitIcon = createIcon(
 	'ExitIcon'
 );
 
-export const ChevronsLeftIcon = createIcon(
+export const FactoryIcon = createIcon(
 	<Fragment>
-		<path d="M18.5 7.5L14.1179 11.8821C14.0528 11.9472 14.0528 12.0528 14.1179 12.1179L18.5 16.5" />
-		<path d="M9.5 7.5L5.11785 11.8821C5.05276 11.9472 5.05276 12.0528 5.11785 12.1179L9.5 16.5" />
+		<path d="M15.7575 12.4394L17.7575 12.9394C18.3886 13.0972 19 12.6198 19 11.9692V3C19 2.44772 18.5523 2 18 2H16C15.4477 2 15 2.44772 15 3V11.4692C15 11.9281 15.3123 12.3281 15.7575 12.4394ZM3.24254 9.31063L21.2425 13.8106C21.6877 13.9219 22 14.3219 22 14.7808V21C22 21.5523 21.5523 22 21 22H3C2.44772 22 2 21.5523 2 21V10.2808C2 9.63021 2.61139 9.15285 3.24254 9.31063Z" />
+		<rect x="7" y="18" width="1" height="1" transform="rotate(180 7 18)" />
+		<rect x="11" y="17" width="1" height="1" />
 	</Fragment>,
-	'ChevronsLeftIcon'
+	'FactoryIcon'
+);
+
+export const HomeIcon = createIcon(
+	<Fragment>
+		<path d="M20 11V21C20 21.5523 19.5523 22 19 22H5C4.44772 22 4 21.5523 4 21V11" />
+		<path d="M2 12C3.83629 10.0774 12 2 12 2C12 2 20.1637 10.0774 22 12" />
+	</Fragment>,
+	'HomeIcon'
 );
