--- packages/react/src/icon/icons/ChartPieIcon.tsx
@@ -1,12 +1,0 @@
-import { Fragment } from 'react';
-import { createIcon } from '../Icon';
-
-export const ChartPieIcon = createIcon(
-	<Fragment>
-		<path d="M12 22C17.5228 22 22 17.5228 22 12C22 6.47715 17.5228 2 12 2C6.47715 2 2 6.47715 2 12C2 17.5228 6.47715 22 12 22Z" />
-		<path d="M7 4L12 12" />
-		<path d="M17.4545 20.1818L12 12" />
-		<path d="M22 12H12" />
-	</Fragment>,
-	'ChartPieIcon'
-);
