--- packages/react/src/icon/icons/HomeIcon.tsx
@@ -1,10 +1,0 @@
-import { Fragment } from 'react';
-import { createIcon } from '../Icon';
-
-export const HomeIcon = createIcon(
-	<Fragment>
-		<path d="M20 11V21C20 21.5523 19.5523 22 19 22H5C4.44772 22 4 21.5523 4 21V11" />
-		<path d="M2 12C3.83629 10.0774 12 2 12 2C12 2 20.1637 10.0774 22 12" />
-	</Fragment>,
-	'HomeIcon'
-);
