+++ packages/react/src/icon/index.ts
@@ -14,7 +14,6 @@ export { AvatarIcon } from './icons/AvatarIcon';
 export { CalendarIcon } from './icons/CalendarIcon';
 export { ChartBarIcon } from './icons/ChartBarIcon';
 export { ChartLineIcon } from './icons/ChartLineIcon';
-export { ChartPieIcon } from './icons/ChartPieIcon';
 export { ChevronUpIcon } from './icons/ChevronUpIcon';
 export { ChevronDownIcon } from './icons/ChevronDownIcon';
 export { ChevronLeftIcon } from './icons/ChevronLeftIcon';
@@ -27,10 +26,8 @@ export { DownloadIcon } from './icons/DownloadIcon';
 export { EditIcon } from './icons/EditIcon';
 export { EmailIcon } from './icons/EmailIcon';
 export { FacebookIcon } from './icons/FacebookIcon';
-export { FactoryIcon } from './icons/FactoryIcon';
 export { FilterIcon } from './icons/FilterIcon';
 export { HelpIcon } from './icons/HelpIcon';
-export { HomeIcon } from './icons/HomeIcon';
 export { InfoIcon } from './icons/InfoIcon';
 export { InfoFilledIcon } from './icons/InfoFilledIcon';
 export { InstagramIcon } from './icons/InstagramIcon';
